Split HTML review renderer tests
