+++ docs/DECISIONS.md
@@ -317,3 +317,4 @@
 - HTML renderer architecture gates live in `ParityHTMLGateTests`, not the broad `ParityGateTests` catch-all. Pure HTML renderer delegation checks for tool cards, top bar, terminal, secondary panes, review, transcript, and sidebar should stay together; browser-specific HTML rendering stays in `ParityBrowserGateTests`, and mixed native/composer/workspace surface gates remain in the broad suite until they have enough focused domain coverage to split cleanly.
 - HTML tool-card renderer coverage lives in `WorkspaceHTMLToolCardRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Tool-card output, approval actions, artifact chips, image/document/appshot previews, and transcript ordering should stay beside the focused renderer so artifact UI work does not inflate the generic workspace surface suite.
 - HTML terminal renderer coverage lives in `WorkspaceHTMLTerminalRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Visible terminal pane markup, command rows, clear controls, and running/stopped status labels should stay beside terminal HTML so terminal UI work can evolve without expanding generic surface projection tests.
+- HTML review renderer coverage lives in `WorkspaceHTMLReviewRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Static review-pane markup, review file/action/hunk/line/comment rendering, and review action data attributes should stay beside review HTML while review-surface data projection remains in focused workspace surface tests until it has a separate builder suite.
